Moved scripts around to improve structure
