+++ Assets/Scripts/ShopTip.cs
