+++ Assets/Scripts/System/Paths.cs.meta
